Try to simplify AppleWatch.swift
WatchConnectivity part of TheGreatKit was a total mess.
It had several classes, it was hard to reason about, and all
these classes didn't have clear, visible responsibilities. This
update aims to simplify it a little bit, removing some of the
confusing Alba (reactive) parts, embracing RegistryDescriptor
types and making some renamings to make the code cleaner.

Still not happy with the results, but it's better now.